Update cleanup.sh
Found an issue where instead of outputting .webp to the image: line in the YAML frontmatter the script was outputting .webp). This should be fixed.

I also wanted the steps commented out in the script to be a bit more clearly defined for future reference and configuration changes.

I also decided to change the name of the directory in my 5eVault from [2. Mechanics] to [2. Lexicon] because I found that more accurate to the purpose of the directory.